+++ test/preprocess4/edge-case/advanced-binding-element.js
@@ -10,7 +10,7 @@ import {
   html,
   i18n,
   bind
-} from '../../../i18n.js';
+} from '../../../i18n-core.js';
 import '@polymer/iron-input/iron-input.js';
 import deepcopy from 'deepcopy/dist/deepcopy.js';
 switch (syntax) {