+++ test/preprocess4/edge-case/complex-compound-binding-element.js
@@ -11,7 +11,7 @@ import {
   html,
   i18n,
   bind
-} from '../../../i18n.js';
+} from '../../../i18n-core.js';
 switch (syntax) {
 default:
 case 'element-binding': {