+++ test/preprocess4/lit-clock/clock.js
@@ -20,7 +20,7 @@ import {
   html,
   i18n,
   bind
-} from '../../../i18n.js';
+} from '../../../i18n-core.js';
 import {
   getMessage,
   binding as messageBinding