+++ test/preprocess4/minimal-test-imports.js
@@ -4,7 +4,7 @@ Copyright (c) 2019, Tetsuya Mori <t2y3141592@gmail.com>. All rights reserved.
 */
 import './fake-server.js';
 import './test-runner.js';
-import '../../i18n.js';
+import '../../i18n-core.js';
 import './lit-clock/clock.js';
 import './simple-text-element/simple-text-element.js';
 import './minimal-test-suites.js';