+++ test/preprocess4/multiple-case/multiple-element.js
@@ -8,7 +8,7 @@ import {
   html,
   i18n,
   bind
-} from '../../../i18n.js';
+} from '../../../i18n-core.js';
 import './item-element.js';
 switch (syntax) {
 default: