+++ test/preprocess4/simple-attribute-element/text-attribute-element.js
@@ -11,7 +11,7 @@ import {
   html,
   i18n,
   bind
-} from '../../../i18n.js';
+} from '../../../i18n-core.js';
 import deepcopy from 'deepcopy/dist/deepcopy.js';
 const $_documentContainer = document.createElement('template');
 $_documentContainer.innerHTML = `<i18n-attr-repo>